requirements.txt
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-html>=4.1.0
python-dotenv>=1.0.0
